modify PostCard to DefaultPostCard and create PostCard
