+++ feature/src/main/java/com/example/feature/screens/hub/setting/my_posts/MyPosts.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.domain.presentation.navigation.Navigation
 import com.example.feature.R
-import com.example.feature.components.card.postcard.PostCard
+import com.example.feature.components.card.postcard.DefaultPostCard
 import com.example.feature.components.card.postcard.PostCardAction
 import com.example.feature.components.card.postcard.PostCardWithDetailButton
 import com.example.feature.components.indicator.LoadingUI
@@ -184,7 +184,7 @@ private fun Likes(
                 myPostsState.likedPosts.size,
                 key = { index: Int -> myPostsState.likedPosts[index].id },
             ) { i ->
-                PostCard(
+                DefaultPostCard(
                     post = myPostsState.likedPosts[i],
                     hubState = hubState,
                     isIncrementView = false,
@@ -239,7 +239,7 @@ private fun Reposts(
                 myPostsState.repostedPosts.size,
                 key = { index: Int -> myPostsState.repostedPosts[index].id },
             ) { i ->
-                PostCard(
+                DefaultPostCard(
                     post = myPostsState.repostedPosts[i],
                     hubState = hubState,
                     isIncrementView = false,
